+++ lib/src/fx_state/fx_value.dart
@@ -1,23 +1,23 @@
 import 'dart:math';
 
 import 'package:flutter/widgets.dart';
+import 'package:flutter_fx/src/fx_state/fx_widgets.dart';
 
 import 'fx_notifier.dart';
-import 'fx_state.dart';
 
 typedef FxString = Fx<String>;
 typedef FxBool = Fx<bool>;
 typedef FxList<T> = Fx<List<T>>;
 typedef FxMap<T, M> = Fx<Map<T, M>>;
 
-
 /// Extension to convert a value to a `Fx` instance.
 extension FxValue<T> on T {
   /// Returns a `Fx` instance with [this] `T` as initial value.
   Fx<T> get toFx => Fx<T>(this);
-  
-  /// Returns a `Fx` instance with `null` of `T` as initial value.
-  Fx<T?> get toFxNullable => (null as T?).toFx;
+}
+
+class FxNullable {
+  static Fx<T> setNull<T>() => (null as T).toFx;
 }
 
 class Fx<T extends dynamic> {
@@ -28,7 +28,7 @@ class Fx<T extends dynamic> {
   Fx(this._value);
 
   set value(value) {
-    _value = value;    
+    _value = value;
     FxStateNotifier.instance.notifyBuilders(fxIdentifier);
   }